Address review feedback: merge multi_database.go into dsn.go, drop index_properties
Per Cristian's review on the postgres PR:
- Move discoverDatabases/connectToDatabase and selectAllDatabases into
  dsn.go, since DSN parsing and per-database connection fan-out go
  together; drop the now-empty multi_database.go/_test.go.
- Drop the pg_index_properties metric and its is_primary column/join:
  schema_details already surfaces primary-key info, and index_stats has
  no equivalent in the mysql PR. This also removes the double-append
  needed only to build that metric's label list.
- Trim the now-stale "field names match upstream" comment on index_stats.
