+++ internal/component/database_observability/postgres/collector/schema_details.go
@@ -26,14 +26,6 @@ const (
 )
 
 const (
-	// selectAllDatabases makes use of the initial DB connection to discover other databases on the same Postgres instance
-	selectAllDatabases = `
-		SELECT datname
-		FROM pg_database
-		WHERE datistemplate = false
-			AND has_database_privilege(datname, 'CONNECT')
-			AND datname NOT IN %s`
-
 	// selectSchemaNames gets all user-defined schemas, excluding system schemas
 	selectSchemaNames = `
 	SELECT
